--- .gitattributes
@@ -1,1 +1,0 @@
-*.pdf filter=lfs diff=lfs merge=lfs -text
