--- .gitignore
@@ -1,6 +1,0 @@
-DIC_irgendwie/Debug/*
-*.o
-*.elf
-*.hex
-*.lss
-*.srec
