--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "CAN_LIB"]
-	path = CAN_LIB
-	url = http://10.74.3.17/git/Electronics/CAN_LIB.git
