--- DS_OLED_EA.pdf
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:89860d7282e1139e7335cc1133307ed78a16d46f7d537e77136840fdd11be4f1
-size 1306280
