--- dic/dic/dic.atsln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Atmel Studio Solution File, Format Version 11.00
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{54F91283-7BC4-4236-8FF9-10F437C3AD48}") = "dic", "dic\dic.cproj", "{DCE6C7E3-EE26-4D79-826B-08594B9AD897}"
-EndProject
-Project("{54F91283-7BC4-4236-8FF9-10F437C3AD48}") = "test-display", "test-display\test-display.cproj", "{4FBD17E9-3AB8-4935-84B9-DC9346007D17}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|AVR = Debug|AVR
-		Release|AVR = Release|AVR
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{DCE6C7E3-EE26-4D79-826B-08594B9AD897}.Debug|AVR.ActiveCfg = Debug|AVR
-		{DCE6C7E3-EE26-4D79-826B-08594B9AD897}.Debug|AVR.Build.0 = Debug|AVR
-		{DCE6C7E3-EE26-4D79-826B-08594B9AD897}.Release|AVR.ActiveCfg = Release|AVR
-		{DCE6C7E3-EE26-4D79-826B-08594B9AD897}.Release|AVR.Build.0 = Release|AVR
-		{4FBD17E9-3AB8-4935-84B9-DC9346007D17}.Debug|AVR.ActiveCfg = Debug|AVR
-		{4FBD17E9-3AB8-4935-84B9-DC9346007D17}.Debug|AVR.Build.0 = Debug|AVR
-		{4FBD17E9-3AB8-4935-84B9-DC9346007D17}.Release|AVR.ActiveCfg = Release|AVR
-		{4FBD17E9-3AB8-4935-84B9-DC9346007D17}.Release|AVR.Build.0 = Release|AVR
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
