--- dic/dic/test-display/Debug/makedep.mk
@@ -1,6 +1,0 @@
-################################################################################
-# Automatically-generated file. Do not edit or delete the file
-################################################################################
-
-main.c
-
