--- dic/dic/test-display/Debug/test-display.eep
@@ -1,1 +1,0 @@
-:00000001FF
