--- dic/dic/test-display/Debug/test-display.srec
@@ -1,22 +1,0 @@
-S0140000746573742D646973706C61792E737265632D
-S11300000C944A000C945F000C945F000C945F0005
-S11300100C945F000C945F000C945F000C945F00E0
-S11300200C945F000C945F000C945F000C945F00D0
-S11300300C945F000C945F000C945F000C945F00C0
-S11300400C945F000C945F000C945F000C945F00B0
-S11300500C945F000C945F000C945F000C945F00A0
-S11300600C945F000C945F000C945F000C945F0090
-S11300700C945F000C945F000C945F000C945F0080
-S11300800C945F000C945F000C945F000C945F0070
-S11300900C945F0011241FBECFEFD0E1DEBFCDBFB3
-S11300A011E0A0E0B1E0E8E2F1E002C005900D92B9
-S11300B0A830B107D9F70E9489000C9492000C94DF
-S11300C000008FEF8DB987E087B918B81EB808957E
-S11300D08EB9613011F4409A01C04098429A84E686
-S11300E090E00197F1F7429808950F931F93CF93EF
-S11300F0DF93C0E0D1E007E011E060E089910E9465
-S11301006800C017D107C9F7DF91CF911F910F91F4
-S113011008950E9461000E94750061E081E40E94DC
-S10B01206800FBCFF894FFCF47
-S10B01283908061700020C005F
-S9030000FC
