--- dic/dic/test-display/test-display.componentinfo.xml
@@ -1,86 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Store xmlns:i="http://www.w3.org/2001/XMLSchema-instance" xmlns="AtmelPackComponentManagement">
-	<ProjectComponents>
-		<ProjectComponent z:Id="i1" xmlns:z="http://schemas.microsoft.com/2003/10/Serialization/">
-			<CApiVersion></CApiVersion>
-			<CBundle></CBundle>
-			<CClass>Device</CClass>
-			<CGroup>Startup</CGroup>
-			<CSub></CSub>
-			<CVariant></CVariant>
-			<CVendor>Atmel</CVendor>
-			<CVersion>1.1.0</CVersion>
-			<DefaultRepoPath>C:/atmelstudio\7.0\Packs</DefaultRepoPath>
-			<DependentComponents xmlns:d4p1="http://schemas.microsoft.com/2003/10/Serialization/Arrays" />
-			<Description></Description>
-			<Files xmlns:d4p1="http://schemas.microsoft.com/2003/10/Serialization/Arrays">
-				<d4p1:anyType i:type="FileInfo">
-					<AbsolutePath>C:/atmelstudio\7.0\Packs\atmel\ATmega_DFP\1.1.130\include</AbsolutePath>
-					<Attribute></Attribute>
-					<Category>include</Category>
-					<Condition>C</Condition>
-					<FileContentHash i:nil="true" />
-					<FileVersion></FileVersion>
-					<Name>include</Name>
-					<SelectString></SelectString>
-					<SourcePath></SourcePath>
-				</d4p1:anyType>
-				<d4p1:anyType i:type="FileInfo">
-					<AbsolutePath>C:/atmelstudio\7.0\Packs\atmel\ATmega_DFP\1.1.130\include\avr\iocan64.h</AbsolutePath>
-					<Attribute></Attribute>
-					<Category>header</Category>
-					<Condition>C</Condition>
-					<FileContentHash>tCM3rmVgoF4kPkNAAQf+pw==</FileContentHash>
-					<FileVersion></FileVersion>
-					<Name>include/avr/iocan64.h</Name>
-					<SelectString></SelectString>
-					<SourcePath></SourcePath>
-				</d4p1:anyType>
-				<d4p1:anyType i:type="FileInfo">
-					<AbsolutePath>C:/atmelstudio\7.0\Packs\atmel\ATmega_DFP\1.1.130\templates\main.c</AbsolutePath>
-					<Attribute>template</Attribute>
-					<Category>source</Category>
-					<Condition>C Exe</Condition>
-					<FileContentHash>+gxvHOYHEk7FNtEarnkdGQ==</FileContentHash>
-					<FileVersion></FileVersion>
-					<Name>templates/main.c</Name>
-					<SelectString>Main file (.c)</SelectString>
-					<SourcePath></SourcePath>
-				</d4p1:anyType>
-				<d4p1:anyType i:type="FileInfo">
-					<AbsolutePath>C:/atmelstudio\7.0\Packs\atmel\ATmega_DFP\1.1.130\templates\main.cpp</AbsolutePath>
-					<Attribute>template</Attribute>
-					<Category>source</Category>
-					<Condition>C Exe</Condition>
-					<FileContentHash>YXFphlh0CtZJU+ebktABgQ==</FileContentHash>
-					<FileVersion></FileVersion>
-					<Name>templates/main.cpp</Name>
-					<SelectString>Main file (.cpp)</SelectString>
-					<SourcePath></SourcePath>
-				</d4p1:anyType>
-				<d4p1:anyType i:type="FileInfo">
-					<AbsolutePath>C:/atmelstudio\7.0\Packs\atmel\ATmega_DFP\1.1.130\gcc\dev\at90can64</AbsolutePath>
-					<Attribute></Attribute>
-					<Category>libraryPrefix</Category>
-					<Condition>GCC</Condition>
-					<FileContentHash i:nil="true" />
-					<FileVersion></FileVersion>
-					<Name>gcc/dev/at90can64</Name>
-					<SelectString></SelectString>
-					<SourcePath></SourcePath>
-				</d4p1:anyType>
-			</Files>
-			<PackName>ATmega_DFP</PackName>
-			<PackPath>C:/atmelstudio/7.0/Packs/atmel/ATmega_DFP/1.1.130/Atmel.ATmega_DFP.pdsc</PackPath>
-			<PackVersion>1.1.130</PackVersion>
-			<PresentInProject>true</PresentInProject>
-			<ReferenceConditionId>AT90CAN64</ReferenceConditionId>
-			<RteComponents xmlns:d4p1="http://schemas.microsoft.com/2003/10/Serialization/Arrays">
-				<d4p1:string></d4p1:string>
-			</RteComponents>
-			<Status>Resolved</Status>
-			<VersionMode>Fixed</VersionMode>
-			<IsComponentInAtProject>true</IsComponentInAtProject>
-		</ProjectComponent>
-	</ProjectComponents>
-</Store>
