--- dsplib.h
@@ -1,10 +1,0 @@
-//
-// Created by sbiewald on 17.11.18.
-//
-
-#ifndef TY19_DIC_DSPLIB_H
-#define TY19_DIC_DSPLIB_H
-
-
-
-#endif //TY19_DIC_DSPLIB_H
