--- errors.c
@@ -1,1 +1,0 @@
-extern unsigned long system_tick()
